requirements.txt
PyYAML>=6.0
markdown>=3.0
